admin recipe and report

+++ src/components/Admin/Recipe/index.js
@@ -2,12 +2,18 @@ import React, { useEffect, useState } from "react";
 import adminApi from "../../../api/adminApi";
 import userPageApi from "../../../api/userPageApi";
 import { getToken } from "../../../features/sessionStorage";
-import { Link } from "react-router-dom";
-const Recipe = () => {
+import { Link, useOutletContext } from "react-router-dom";
+import { Modal } from "react-bootstrap";
+import "./../main.scss";
+
+const Recipe = ({ parentCallback, index }) => {
   const token = getToken();
   const [data, setData] = useState();
+  const [urlPage, setUrlPage] = useOutletContext();
 
   useEffect(() => {
+    setUrlPage("recipe");
+
     adminApi
       .getWaitRecipe(getToken())
       .then((res) => {
@@ -36,29 +42,33 @@ const Recipe = () => {
   }
 
   const TableComponent = () => {
-    console.log(data?.length)
+    console.log(data?.length);
     return (
       <tbody>
-      {
-        data?.length ?
+        {data?.length ? (
           data.map((value, index) => {
             return <RowComponent value={value} index={index} key={index} />;
-          }) : (
-            <tr>
-              <td colSpan={5}>Không có sản phẩm nào</td>
-            </tr>
-          )
-        }
+          })
+        ) : (
+          <tr>
+            <td colSpan={5}>Không có sản phẩm nào</td>
+          </tr>
+        )}
       </tbody>
     );
   };
 
   const RowComponent = ({ value, index }) => {
-    const [isLoading, setIsLoading] = useState();
+    const [isLoadingDelete, setIsLoadingDelete] = useState();
+    const [isLoadingPublish, setIsLoadingPublish] = useState();
+    const [isLoadingReject, setIsLoadingReject] = useState();
     const [isDelete, setIsDelete] = useState();
     const [isPublish, setIsPublish] = useState();
+    const [isReject, setIsReject] = useState();
     const [isDisable, setIsDisable] = useState();
     const [owner, setOwner] = useState();
+    const [show, setShow] = useState();
+    let feedback = '';
 
     useEffect(() => {
       userPageApi
@@ -73,9 +83,13 @@ const Recipe = () => {
         });
     }, []);
 
+    useEffect(() => {
+      console.log(feedback);
+    });
+
     const handlePublish = (recipeId) => {
       if (isPublish || isDisable) return;
-      setIsLoading(true);
+      setIsLoadingPublish(true);
       adminApi
         .publishRecipe(getToken(), recipeId)
         .then((res) => {
@@ -83,19 +97,43 @@ const Recipe = () => {
           console.log(res);
           if (res.data.messageCode === 1) {
             setIsDisable(true);
-            setIsPublish(true)
+            setIsPublish(true);
           }
-          setIsLoading(false);
+          setIsLoadingPublish(false);
         })
         .catch((err) => {
           console.log(err);
-          setIsLoading(false);
+          setIsLoadingPublish(false);
+        });
+    };
+
+    const handleReject = (recipeId, reason) => {
+      if (isReject || isDisable) return;
+      setIsLoadingReject(true);
+      adminApi
+        .rejectRecipe(getToken(), recipeId, reason)
+        .then((res) => {
+          console.log(res);
+          if (res.data.messageCode === 1) {
+            setIsDisable(true);
+            setIsReject(true);
+          }
+          setIsLoadingReject(false);
+        })
+        .catch((err) => {
+          console.log(err);
+          setIsLoadingReject(false);
+          setIsReject(false);
         });
     };
 
     const deleteRecipe = (recipeId) => {
       if (isDelete || isDisable) return;
-      setIsLoading(true);
+      setIsLoadingDelete(true);
+      if (!confirm("Bạn có muốn xoá công thức: " + value?.title)) {
+        setIsLoadingDelete(false);
+        return;
+      }
       adminApi
         .deleteRecipe(getToken(), recipeId)
         .then((res) => {
@@ -103,25 +141,142 @@ const Recipe = () => {
           console.log(res);
           if (res.data.messageCode === 1) {
             setIsDisable(true);
-            setIsDelete(true)
+            setIsDelete(true);
           }
-          setIsLoading(false);
+          setIsLoadingDelete(false);
         })
         .catch((err) => {
           console.log(err);
-          setIsLoading(false);
+          setIsLoadingDelete(false);
         });
     };
 
-    return (
-      <tr key={index} data-index={index} className={isDisable && 'opacity-50'}>
-        <td style={{ textTransform: "capitalize" }}>
-          {isDisable ? (
-            <span className="recipe-link">{value?.title}</span>
-          ) : (
-            <Link to={`/recipe/${value?.id}`} className='recipe-link'>{value?.title}</Link>
+    function ModalActive({ children, title }) {
+      const handleClose = () => setShow(false);
+      return (
+        <Modal
+          show={show}
+          onHide={handleClose}
+          centered
+          dialogClassName="rounded-3rem mw-100"
+          animation={false}
+        >
+          <Modal.Header closeButton className="pe-4">
+            <h5 className="text-center flex-grow-1 m-0">
+              <span className="text-info fw-bolder"> Demo: </span>{" "}
+              {`  ${title}`}
+            </h5>
+          </Modal.Header>
+          <Modal.Body className="shadow">{children}</Modal.Body>
+        </Modal>
+      );
+    }
+
+    const RejectComponent = () => {
+      const [input, setInput] = useState(feedback);
+
+      useEffect(() => {
+        feedback = input;
+      })
+
+      return (
+        <div className="recipe-feedback">
+          <h3>Góp ý</h3>
+          { !isReject && (
+            <textarea
+              placeholder="Điền góp ý hoặc lý do từ chối?"
+              className='recipe-feedback__input'
+              value={input}
+              onChange={(event) => setInput(event.target.value)}
+            ></textarea>
           )
           }
+          <div className="d-flex justify-content-evenly flex-wrap flex-column">
+            <button
+              className={
+                "recipe-button recipe-button--warning" +
+                (isDisable ? " active" : "")
+              }
+              onClick={(event) => handleReject(value?.id, feedback)}
+            >
+              {isLoadingReject ? (
+                <React.Fragment>
+                  <div
+                    className="spinner-border text-light"
+                    style={{ width: "1rem", height: "1rem" }}
+                    role="status"
+                  >
+                    <span className="visually-hidden">Loading...</span>
+                  </div>
+                  <span> Loading...</span>
+                </React.Fragment>
+              ) : (
+                <React.Fragment>
+                  {isReject ? "Đã từ chôi" : "Từ chối"}
+                </React.Fragment>
+              )}
+            </button>
+            <button
+              className={
+                "recipe-button recipe-button--delete" +
+                (isDisable ? " active" : "")
+              }
+              onClick={(event) => deleteRecipe(value?.id)}
+            >
+              {isLoadingDelete ? (
+                <React.Fragment>
+                  <div
+                    className="spinner-border text-light"
+                    style={{ width: "1rem", height: "1rem" }}
+                    role="status"
+                  >
+                    <span className="visually-hidden">Loading...</span>
+                  </div>
+                  <span> Loading...</span>
+                </React.Fragment>
+              ) : (
+                <React.Fragment>{isDelete ? "Đã Xoá" : "Xoá"}</React.Fragment>
+              )}
+            </button>
+            <button
+              className={
+                "recipe-button recipe-button--publish" +
+                (isDisable ? " active" : "")
+              }
+              onClick={(event) => handlePublish(value?.id)}
+            >
+              {isLoadingPublish ? (
+                <React.Fragment>
+                  <div
+                    className="spinner-border text-light"
+                    style={{ width: "1rem", height: "1rem" }}
+                    role="status"
+                  >
+                    <span className="visually-hidden">Loading...</span>
+                  </div>
+                  <span> Loading...</span>
+                </React.Fragment>
+              ) : (
+                <React.Fragment>
+                  {isPublish ? "Đã Đăng" : "Đăng"}
+                </React.Fragment>
+              )}
+            </button>
+          </div>
+        </div>
+      );
+    };
+
+    return (
+      <tr key={index} data-index={index} className={isDisable && "opacity-50"}>
+        <td style={{ textTransform: "capitalize" }}>
+          {isDisable ? (
+            <span className="section-link">{value?.title}</span>
+          ) : (
+            <Link to={`/recipe/${value?.id}`} className="section-link">
+              {value?.title}
+            </Link>
+          )}
         </td>
         <td style={{ textAlign: "center" }}>{value?.cooking_time}</td>
         <td style={{ textAlign: "center" }}>{value?.amount_of_people}</td>
@@ -132,10 +287,48 @@ const Recipe = () => {
         </td>
         <td style={{ textAlign: "center" }}>
           <button
-            className={"recipe-button" + (isPublish ? " active" : "")}
+            className={
+              "section-button section-button--info" +
+              (isPublish ? " active" : "")
+            }
+            onClick={(event) => setShow(!show)}
+          >
+            Demo
+          </button>
+          <ModalActive title={value?.title}>
+            <div className="container-fluid">
+              <div className="row">
+                <div className="col-12 col-md-8 col-lg-9 col-xl-10">
+                  <iframe
+                    src={`/recipe/${value?.id}`}
+                    title="công thức"
+                    style={{ width: "100%", height: "85vh" }}
+                  ></iframe>
+                </div>
+                <div className="col-12 col-md-4 col-lg-3 col-xl-2">
+                  <RejectComponent />
+                </div>
+              </div>
+            </div>
+          </ModalActive>
+        </td>
+        <td style={{ textAlign: "center" }}>
+          <Link
+            className={"section-button" + (isPublish ? " active" : "")}
+            to={`/update-recipe/${value?.id}`}
+          >
+            Sửa
+          </Link>
+        </td>
+        <td style={{ textAlign: "center" }}>
+          <button
+            className={
+              "section-button section-button--publish" +
+              (isPublish ? " active" : "")
+            }
             onClick={(event) => handlePublish(value?.id)}
           >
-            {isLoading ? (
+            {isLoadingPublish ? (
               <React.Fragment>
                 <div
                   className="spinner-border text-light"
@@ -154,11 +347,12 @@ const Recipe = () => {
         <td style={{ textAlign: "center" }}>
           <button
             className={
-              "recipe-button recipe-button--delete" + (isDelete ? " active" : "")
+              "section-button section-button--delete" +
+              (isDelete ? " active" : "")
             }
             onClick={(event) => deleteRecipe(value?.id)}
           >
-            {isLoading ? (
+            {isLoadingDelete ? (
               <React.Fragment>
                 <div
                   className="spinner-border text-light"
@@ -179,7 +373,7 @@ const Recipe = () => {
   };
 
   return (
-    <div className="recipe-box">
+    <div className="section-box">
       <div className="recent_project">
         <div className="card_header">
           <h2>Công thức mới</h2>
@@ -191,8 +385,10 @@ const Recipe = () => {
               <td style={{ textAlign: "center" }}>Thời gian nấu</td>
               <td style={{ textAlign: "center" }}>Số lượng người ăn</td>
               <td style={{ textAlign: "center" }}>Đóng góp</td>
-              <td style={{ textAlign: "center" }}>Xoá</td>
+              <td style={{ textAlign: "center" }}>Demo</td>
+              <td style={{ textAlign: "center" }}>Sửa</td>
               <td style={{ textAlign: "center" }}>Duyệt</td>
+              <td style={{ textAlign: "center" }}>Xoá</td>
             </tr>
           </thead>
           <TableComponent />
